requirements.txt
playwright
numpy
pandas
streamlit
openai
python-dotenv